Extract test command and shared options from cli/index.ts
