+++ cli/commands/index.ts
@@ -3,3 +3,4 @@ export { helpCommand } from "df/cli/commands/help_command";
 export { initCommand } from "df/cli/commands/init_command";
 export { initCredsCommand } from "df/cli/commands/init_creds_command";
 export { installCommand } from "df/cli/commands/install_command";
+export { testCommand } from "df/cli/commands/test_command";
